forcing a wakelock release from the sensor.qcom service
forcing a wakelock release from the sensor.qcom service

As we don't yet have a replacement for java's service manager (which is
the one responsible for starting services per requests), we need to
manually start the sensorservice service during boot. In mako's case,
this will result in a wakelock acquire by the sensor.qcom service, which
blocks deep suspend. As a workaround, a script is used just after the
sensorservice is up, which manually releases the wakelock, for it to
sleep properly and have a long battery life.

Change-Id: Ib43d2ca0a64aee3025edf35f3ec4e02601755609
Signed-off-by: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com>